Fix `make all`
This change makes `make publish-go` work again. This is because `git
subtree split` will create different commits each time, since they each
commit will contain the current user's username / email as the committer
and the current timestamp as the commit time, therefore changing the
hash of the commits.

Instead, this will now create a new worktree so that the original
contents of the protocol-go repository will be rebased on to preserve
whatever commits were there before.